(#99) implement marking notifications as done on github
and again, I had to change the interface to accept a long instead of
an int.

+++ include/ghcli/github/status.h
@@ -33,6 +33,6 @@
 #include <ghcli/status.h>
 
 size_t github_get_notifications(ghcli_notification **notifications, int count);
-void   github_notification_mark_as_read(int id);
+void   github_notification_mark_as_read(long id);
 
 #endif /* GITHUB_STATUS_H */
